project/requirements.txt
ansible==2.9.7
acitoolkit==0.4
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
deepdiff==4.3.2
dominate==2.5.1
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-HTTPAuth==4.0.0
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
gitdb==4.0.5
GitPython==3.1.2
graphviz==0.14
idna==2.9
importlib-metadata==1.6.0
ipaddress==1.0.23
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
ordered-set==4.0.1
py-radix==0.10.0
PyMySQL==0.9.3
pyrsistent==0.16.0
PyYAML==5.3.1
requests==2.23.0
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.17
tabulate==0.8.7
urllib3==1.25.9
visitor==0.1.3
websocket-client==0.57.0
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
